+++ Release_Notes.md
@@ -11,6 +11,9 @@
 - Forgot password's cancel now closes the modal and opens a new one for login
 - Moved version number to GrowthWare.Framework.csproj
     - Update both launch.json and GrowthWare.Framework.csproj because launch.json still works for the DatabaseManager.
+- Added ADatabaseTable and IDatabaseTable as a replacement for AbstractDatabaseFunctions and IDatabaseFunctions.
+    - Making some of the methods and properties static
+    - Added addtributes to the properties giving clairty to them
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure